+++ tests/connectors/test_connectors.py
@@ -1,6 +1,7 @@
 # _*_ coding: utf-8 _*_
 import pytest
 from elasticsearch import Elasticsearch
+from elasticsearch import AsyncElasticsearch
 
 from fhirpath.connectors import create_connection
 from fhirpath.connectors.factory import pg
@@ -15,8 +16,16 @@ def test_es_connection_creation(es):
     host, port = es
     conn_str = "es://@{0}:{1}/".format(host, port)
     conn = create_connection(conn_str, Elasticsearch)
-
     assert conn.raw_connection.ping() is True
+
+
+async def test_async_es_connection_creation(es):
+    """ """
+    host, port = es
+    conn_str = "es://@{0}:{1}/".format(host, port)
+    conn = create_connection(conn_str, "elasticsearch.AsyncElasticsearch")
+    assert isinstance(conn.raw_connection, AsyncElasticsearch)
+    assert await conn.raw_connection.ping() is True
 
 
 @pytest.mark.skipif(IS_TRAVIS, reason="ignore for travis environment")